Let's split this up.
